project_Luxonis/python/requirements.txt
Scrapy==2.9.0
requests==2.30.0
requests-file==1.5.1
mysql-connector-python
